% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epo.R
\name{epo}
\alias{epo}
\alias{epo.default}
\alias{epo.tbl}
\alias{epo.xts}
\alias{epo.matrix}
\title{Enhanced Portfolio Optimization (EPO)}
\usage{
epo(x, signal, method = c("simple", "anchored"), w, anchor = NULL)

\method{epo}{default}(x, signal, method = c("simple", "anchored"), w, anchor = NULL)

\method{epo}{tbl}(x, signal, method = c("simple", "anchored"), w, anchor = NULL)

\method{epo}{xts}(x, signal, method = c("simple", "anchored"), w, anchor = NULL)

\method{epo}{matrix}(x, signal, method = c("simple", "anchored"), w, anchor = NULL)
}
\arguments{
\item{x}{A data-set with asset returns. It should be a \code{tibble}, a \code{xts}
or a \code{matrix}.}

\item{signal}{A \code{double} with the investor's belief's (signals, forecasts).}

\item{method}{A \code{character}. One of: \code{simple} or \code{anchored}.}

\item{w}{A \code{double} between \code{0} and \code{1}. The shrinkage level
increases from 0 to 1.}

\item{anchor}{A \code{double} with the investor anchor (benchmark), in which
the allocation should not deviate too much from. Only used when \code{method = "anchored"}.}
}
\value{
The optimal allocation vector.
}
\description{
Computes the optimal portfolio allocation using the EPO method with full-investment
constraint.
}
\examples{
x <- diff(log(EuStockMarkets)) # stock returns
s <- colMeans(x) # it could be any signal

##################
### Simple EPO ###
##################

# Traditional Mean-Variance Analysis
epo(x = x, signal = s, method = "simple", w = 0)

# 100\% Shrinkage
epo(x = x, signal = s, method = "simple", w = 1)

# 50\% Classical MVO and 50\% Shrinkage
epo(x = x, signal = s, method = "simple", w = 0.5)

####################
### Anchored EPO ###
####################

benchmark <- rep(0.25, 4) # 1/N Portfolio

# Traditional Mean-Variance Analysis
epo(x = x, signal = s, method = "anchored", w = 0.0, anchor = benchmark)

# 100\% on the Anchor portfolio
epo(x = x, signal = s, method = "anchored", w = 1.0, anchor = benchmark)

# Somewhere between the two worlds
epo(x = x, signal = s, method = "anchored", w = 0.5, anchor = benchmark)
}
